% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_rule.R
\name{suggest_cond_rule}
\alias{suggest_cond_rule}
\title{Suggest a check for completeness.}
\usage{
suggest_cond_rule(d, vars = names(d))
}
\arguments{
\item{d}{\code{data.frame}, used to generate the checks}

\item{vars}{\code{character} optionally the subset of variables to be used.}
}
\description{
Suggest a check for completeness.
}
\examples{
data(retailers, package="validate")

# will generate check for all columns in retailers that are
# complete.
suggest_na_check(retailers)
}
