% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_check.R
\name{write_type_check}
\alias{write_type_check}
\alias{suggest_type_check}
\title{suggest type check}
\usage{
write_type_check(d, vars = names(d), file = stdout())

suggest_type_check(d, vars = names(d))
}
\arguments{
\item{d}{\code{data.frame}, used to generate the checks}

\item{vars}{\code{character} optionally the subset of variables to be used.}

\item{file}{file to which the checks will be written to.}
}
\description{
suggest type check
}
