% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatesuggest-package.R
\docType{package}
\name{validatesuggest-package}
\alias{validatesuggest}
\alias{validatesuggest-package}
\title{validatesuggest: Generate Suggestions for Validation Rules}
\description{
Generate suggestions for validation rules from a reference data set, which can be used as a starting point for domain specific rules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwindj/validatesuggest}
  \item Report bugs at \url{https://github.com/edwindj/validatesuggest/issues}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/0000-0002-6580-4718}{ORCID})

}
\keyword{internal}
