% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_check.R
\name{write_na_check}
\alias{write_na_check}
\alias{suggest_na_check}
\title{Suggest an check for completeness.}
\usage{
write_na_check(d, vars = names(d), file = stdout())

suggest_na_check(d, vars = names(d))
}
\arguments{
\item{d}{\code{data.frame}, used to generate the checks}

\item{vars}{\code{character} optionally the subset of variables to be used.}

\item{file}{file to which the checks will be written to.}
}
\description{
Suggest an check for completeness.
}
