% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{task2}
\alias{task2}
\title{task2 dataset}
\format{
\describe{
\item{ID}{ID}
\item{Age}{Age of person}
\item{Married}{Marital status}
\item{Employed}{Employed or not}
\item{Working_hours}{Working hours}
}
}
\usage{
task2
}
\description{
Fictuous test data set from European (ESSnet) project on validation 2017.
}
\references{
European (ESSnet) project on validation 2017
}
\keyword{datasets}
