% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatesuggest-package.R
\docType{package}
\name{validatesuggest-package}
\alias{validatesuggest}
\alias{validatesuggest-package}
\title{validatesuggest: Generate Suggestions for Validation Rules}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://edwindj.github.io/validatesuggest}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
